requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database (SQLite async)
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Pydantic & Settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# HTTP & Forms
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Email (OTP sending via Gmail SMTP)
aiosmtplib==3.0.1
